Shift ColorExtensions.cs to CairoExtensions.cs
